fix(www): decouple cms client from next/headers, so it can be used in pages dir
