+++ apps/www/src/lib/apollo/cms-client.ts
@@ -1,41 +1,16 @@
-import {
-  NextSSRInMemoryCache,
-  NextSSRApolloClient,
-} from '@apollo/experimental-nextjs-app-support/ssr'
-import { HttpLink, NormalizedCacheObject } from '@apollo/client'
+import { NormalizedCacheObject } from '@apollo/client'
+import { NextSSRApolloClient } from '@apollo/experimental-nextjs-app-support/ssr'
+import { getCMSClientBase } from './cms-client-base'
 import { draftMode } from 'next/headers'
 
 /**
  * Get a apollo client to interact with DatoCMS on the server.
  *
+ * Sets Draft Mode automatically based on next/headers. If you want to explicitly set it
+ * or use it in the Pages directory, import from @app/lib/apollo/cms-client-base
+ *
  * @returns ApolloClient to interact with DatoCMS
  */
 export function getCMSClient(): NextSSRApolloClient<NormalizedCacheObject> {
-  if (!process.env.DATO_CMS_API_URL) {
-    throw new Error('Missing DatoCMS API URL')
-  }
-  if (!process.env.DATO_CMS_API_TOKEN) {
-    throw new Error('Missing DatoCMS API token')
-  }
-
-  const headers = {
-    Authorization: process.env.DATO_CMS_API_TOKEN,
-    'X-Exclude-Invalid': 'true',
-  }
-
-  if (process.env.DATO_CMS_ENVIRONMENT) {
-    headers['X-Environment'] = process.env.DATO_CMS_ENVIRONMENT
-  }
-
-  if (process.env.DATO_CMS_INCLUDE_DRAFTS === 'true' || draftMode().isEnabled) {
-    headers['X-Include-Drafts'] = 'true'
-  }
-
-  return new NextSSRApolloClient({
-    cache: new NextSSRInMemoryCache(),
-    link: new HttpLink({
-      uri: process.env.DATO_CMS_API_URL,
-      headers,
-    }),
-  })
+  return getCMSClientBase({ draftMode: draftMode().isEnabled })
 }
